refactor: move meshes to separate module
