+++ src/renderer/mod.rs
@@ -1,4 +1,5 @@
 pub mod ecs_renderer;
+mod meshes;
 pub mod metrics;
 pub mod shader;
 pub mod texture;
